requirements.txt
Flask
Flask-SocketIO
psutil
eventlet
Werkzeug
python-dotenv